+++ pkg/util/unstructured.go
@@ -26,7 +26,7 @@ import (
 	"k8s.io/client-go/discovery"
 	"k8s.io/client-go/rest"
 
-	"github.com/kubestellar/kubestellar/api/v1alpha1"
+	"github.com/kubestellar/kubestellar/api/edge/v1alpha1"
 )
 
 const (
